feat(list): rewrite and merge PaginationDataView and RefreshableDataView
